requirements.txt
accelerate==1.11.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
av==16.0.1
blessed==1.25.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
decorator==5.2.1
decord==0.6.0
dill==0.3.8
docstring_parser==0.17.0
einops==0.8.1
exceptiongroup==1.3.1
fastapi==0.124.4
ffmpy==1.0.0
filelock==3.19.1
fire==0.7.1
flash-attn @ https://github.com/Dao-AILab/flash-attention/releases/download/v2.6.3/flash_attn-2.6.3+cu123torch2.4cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=237ef9c6157db394e1ddde4ba609a21ebb98382377a27041edc09318801a6f24
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
gpustat==1.1.1
gradio==5.45.0
gradio_client==1.13.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llamafactory @ git+https://github.com/hiyouga/LLaMA-Factory.git@ddd7dcc7223e4232f06af417148cd3a161c81e04#egg=llamafactory
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
modelscope==1.33.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
networkx==3.3
nltk==3.9.2
numba==0.63.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
peft==0.17.1
pillow==11.3.0
platformdirs==4.5.1
pooch==1.8.2
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rouge-chinese==1.0.3
ruff==0.14.9
safehttpx==0.1.7
safetensors==0.5.3
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.48.0
shellingham==1.5.4
shtab==1.8.0
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soxr==1.0.0
sse-starlette==3.0.4
starlette==0.50.0
sympy==1.13.1
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.22
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.67.1
transformers==4.57.3
triton==3.0.0
trl==0.9.6
typer==0.20.0
typer-slim==0.20.0
typing_extensions==4.15.0
tyro==0.8.14
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
wandb==0.23.1
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
